Updates + test.
