+++ tests/test_subdomains.py
@@ -1,0 +1,60 @@
+import numpy as np
+from math import floor
+
+from conftest import skipif
+from devito import Grid, TimeFunction, Eq, solve, Operator, SubDomains
+
+pytestmark = skipif(['yask', 'ops'])
+
+
+class TestSubdomains(object):
+    """
+    Class for testing SubDomains
+    """
+
+    def test_iterate_NDomains(self):
+        """
+        Test that a set of subdomains are iterated upon correctly.
+        """
+
+        n_domains = 10
+
+        class Inner(SubDomains):
+            name = 'inner'
+
+            def define(self, dimensions):
+                return {d: ('middle', 0, 0) for d in dimensions}
+
+        bounds_xm = np.zeros((n_domains,), dtype=np.int32)
+        bounds_xM = np.zeros((n_domains,), dtype=np.int32)
+        bounds_ym = np.zeros((n_domains,), dtype=np.int32)
+        bounds_yM = np.zeros((n_domains,), dtype=np.int32)
+
+        for j in range(0, n_domains):
+            bounds_xm[j] = j
+            bounds_xM[j] = n_domains-1-j
+            bounds_ym[j] = floor(j/2)
+            bounds_yM[j] = floor(j/2)
+
+        bounds = (bounds_xm, bounds_xM, bounds_ym, bounds_yM)
+
+        inner_sd = Inner(N=n_domains, bounds=bounds)
+
+        grid = Grid(extent=(10, 10), shape=(10, 10), subdomains=(inner_sd, ))
+
+        f = TimeFunction(name='f', grid=grid, dtype=np.int32)
+        f.data[:] = 0.0
+
+        stencil = Eq(f.forward, solve(Eq(f.dt, 1), f.forward),
+                     subdomain=grid.subdomains['inner'])
+
+        op = Operator(stencil)
+
+        op(time_m=0, time_M=9, dt=1)
+        result = f.data[0]
+
+        expected = np.zeros((10, 10), dtype=np.int32)
+        for j in range(0, n_domains):
+            expected[j, bounds_ym[j]:n_domains-bounds_yM[j]] = 10
+
+        assert((np.array(result) == expected).all())
